+++ src/types/mod.rs
@@ -3,10 +3,8 @@ mod db_x509;
 mod email;
 pub mod jwk;
 mod name;
-mod x509_chain;
 
 pub use data_uri::DataUri;
 pub use db_x509::DbX509;
 pub use email::EmailAddress;
 pub use name::PersonalName;
-pub use x509_chain::X509Chain;
